pyproject.toml
[project]
name = "paddle-ocr"
version = "0.1.0"
description = "PaddleOCR demo project"
requires-python = "==3.10.14"
dependencies = [
    "paddleocr>=3.4.0,<4.0",
    "numpy>=2.0",
    "pillow>=12.0",
    "paddlepaddle-gpu>=3.0.0",
    "opencv-python-headless>=4.13.0.90",
    "playwright>=1.57.0",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "httpx>=0.28.1",
]

[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"
extra-index-url = ["https://www.paddlepaddle.org.cn/packages/stable/cu126/"]
index-strategy = "unsafe-best-match"
environments = ["sys_platform == 'linux'"]
